+++ @kiva/kv-shop/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.12.10](https://github.com/kiva/kv-ui-elements/compare/@kiva/kv-shop@1.12.9...@kiva/kv-shop@1.12.10) (2024-07-12)
+
+**Note:** Version bump only for package @kiva/kv-shop
+
+
+
+
+
+## [1.12.9](https://github.com/kiva/kv-ui-elements/compare/@kiva/kv-shop@1.12.8...@kiva/kv-shop@1.12.9) (2024-07-10)
+
+**Note:** Version bump only for package @kiva/kv-shop
+
+
+
+
+
 ## [1.12.8](https://github.com/kiva/kv-ui-elements/compare/@kiva/kv-shop@1.12.7...@kiva/kv-shop@1.12.8) (2024-07-08)
 
 **Note:** Version bump only for package @kiva/kv-shop